+++ src/utils/options-variable.ts
@@ -1,3 +1,5 @@
+import { EventBus } from '@grafana/data';
+
 import { ALL_VALUE_PARAMETER, NO_VALUE_PARAMETER } from '../constants';
 import { RuntimeVariableWithOptions } from '../types';
 import { selectVariableValues } from './variable';
@@ -11,11 +13,13 @@ export const updateVariableOptions = ({
   value,
   emptyValueEnabled,
   variable,
+  panelEventBus,
 }: {
   previousValues: string[];
   value: string | string[];
   emptyValueEnabled: boolean;
   variable: RuntimeVariableWithOptions;
+  panelEventBus: EventBus;
 }) => {
   const updatedValues = Array.isArray(value) ? value : [value];
 
@@ -27,7 +31,7 @@ export const updateVariableOptions = ({
      * Clear Value
      */
     if (updatedValues.length === 0 && emptyValueEnabled) {
-      selectVariableValues([NO_VALUE_PARAMETER], variable);
+      selectVariableValues([NO_VALUE_PARAMETER], variable, panelEventBus);
       return;
     }
 
@@ -35,12 +39,12 @@ export const updateVariableOptions = ({
      * Select all
      */
     if (updatedValues.length === 0 && variable?.multi && variable.includeAll) {
-      selectVariableValues([ALL_VALUE_PARAMETER], variable);
+      selectVariableValues([ALL_VALUE_PARAMETER], variable, panelEventBus);
       return;
     }
 
     const removedValues = previousValues.filter((value) => !updatedValues.includes(value));
-    selectVariableValues(removedValues, variable);
+    selectVariableValues(removedValues, variable, panelEventBus);
     return;
   }
 
@@ -54,7 +58,8 @@ export const updateVariableOptions = ({
   ) {
     selectVariableValues(
       updatedValues.filter((value) => value !== ALL_VALUE_PARAMETER),
-      variable
+      variable,
+      panelEventBus
     );
     return;
   }
@@ -67,12 +72,12 @@ export const updateVariableOptions = ({
     !previousValues.includes(ALL_VALUE_PARAMETER) &&
     updatedValues.includes(ALL_VALUE_PARAMETER)
   ) {
-    selectVariableValues([ALL_VALUE_PARAMETER], variable);
+    selectVariableValues([ALL_VALUE_PARAMETER], variable, panelEventBus);
     return;
   }
 
   /**
    * Select Values
    */
-  selectVariableValues(updatedValues, variable);
+  selectVariableValues(updatedValues, variable, panelEventBus);
 };